+++ tests/TestBase.cs
@@ -1,21 +1,24 @@
-﻿/* AngelCode bitmap font parsing using C#
- * http://www.cyotek.com/blog/angelcode-bitmap-font-parsing-using-csharp
- *
- * Copyright © 2012-2017 Cyotek Ltd.
- *
- * Licensed under the MIT License. See LICENSE.txt for the full text.
- */
-
- using System;
+using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
+
+// AngelCode bitmap font parsing using C#
+// https://www.cyotek.com/blog/angelcode-bitmap-font-parsing-using-csharp
+
+// Copyright © 2012-2020 Cyotek Ltd.
+
+// This work is licensed under the MIT License.
+// See LICENSE.TXT for the full text
+
+// Found this code useful?
+// https://www.paypal.me/cyotek
 
 namespace Cyotek.Drawing.BitmapFont.Tests
 {
   public abstract class TestBase
   {
-    #region Properties
+    #region Protected Properties
 
     protected string DataPath
     {
@@ -29,320 +32,60 @@ namespace Cyotek.Drawing.BitmapFont.Tests
         BitmapFont font;
 
         font = new BitmapFont
-               {
-                 FamilyName = "Arial",
-                 FontSize = 32,
-                 Padding = new Padding(0, 0, 0, 0),
-                 Unicode = true,
-                 StretchedHeight = 100,
-                 Smoothed = true,
-                 Spacing = new Point(1, 1),
-                 LineHeight = 32,
-                 BaseHeight = 26,
-                 AlphaChannel = 1,
-                 SuperSampling = 1,
-                 TextureSize = new Size(256, 256),
-                 Pages = new[]
+        {
+          FamilyName = "Arial",
+          FontSize = 32,
+          Padding = new Padding(0, 0, 0, 0),
+          Unicode = true,
+          StretchedHeight = 100,
+          Smoothed = true,
+          Spacing = new Point(1, 1),
+          LineHeight = 32,
+          BaseHeight = 26,
+          AlphaChannel = 1,
+          SuperSampling = 1,
+          TextureSize = new Size(256, 256),
+          Pages = new[]
                          {
                            new Page(0, Path.Combine(this.DataPath, "simple_0.png"))
                          },
-                 Kernings = new Dictionary<Kerning, int>
+          Kernings = new Dictionary<Kerning, int>
                             {
                               {
                                 new Kerning('f', 'f', -1), -1
                               }
                             },
-                 Characters = new Dictionary<char, Character>
+          Characters = new Dictionary<char, Character>
                               {
-                                {
-                                  'a', new Character
-                                       {
-                                         Char = 'a',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(0, 26, 14, 15)
-                                       }
-                                },
-                                {
-                                  'b', new Character
-                                       {
-                                         Char = 'b',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(1, 6),
-                                         Bounds = new Rectangle(8, 0, 14, 20)
-                                       }
-                                },
-                                {
-                                  'c', new Character
-                                       {
-                                         Char = 'c',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 14,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(30, 21, 13, 15)
-                                       }
-                                },
-                                {
-                                  'd', new Character
-                                       {
-                                         Char = 'd',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(0, 6),
-                                         Bounds = new Rectangle(23, 0, 14, 20)
-                                       }
-                                },
-                                {
-                                  'e', new Character
-                                       {
-                                         Char = 'e',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(198, 0, 15, 15)
-                                       }
-                                },
-                                {
-                                  'f', new Character
-                                       {
-                                         Char = 'f',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 7,
-                                         Offset = new Point(-1, 6),
-                                         Bounds = new Rectangle(126, 0, 10, 20)
-                                       }
-                                },
-                                {
-                                  'g', new Character
-                                       {
-                                         Char = 'g',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(38, 0, 14, 20)
-                                       }
-                                },
-                                {
-                                  'h', new Character
-                                       {
-                                         Char = 'h',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(1, 6),
-                                         Bounds = new Rectangle(98, 0, 13, 20)
-                                       }
-                                },
-                                {
-                                  'i', new Character
-                                       {
-                                         Char = 'i',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 6,
-                                         Offset = new Point(1, 6),
-                                         Bounds = new Rectangle(146, 0, 4, 20)
-                                       }
-                                },
-                                {
-                                  'j', new Character
-                                       {
-                                         Char = 'j',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 6,
-                                         Offset = new Point(-2, 6),
-                                         Bounds = new Rectangle(0, 0, 7, 25)
-                                       }
-                                },
-                                {
-                                  'k', new Character
-                                       {
-                                         Char = 'k',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 14,
-                                         Offset = new Point(1, 6),
-                                         Bounds = new Rectangle(112, 0, 13, 20)
-                                       }
-                                },
-                                {
-                                  'l', new Character
-                                       {
-                                         Char = 'l',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 6,
-                                         Offset = new Point(1, 6),
-                                         Bounds = new Rectangle(151, 0, 4, 20)
-                                       }
-                                },
-                                {
-                                  'm', new Character
-                                       {
-                                         Char = 'm',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 22,
-                                         Offset = new Point(1, 11),
-                                         Bounds = new Rectangle(156, 0, 20, 15)
-                                       }
-                                },
-                                {
-                                  'n', new Character
-                                       {
-                                         Char = 'n',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(1, 11),
-                                         Bounds = new Rectangle(44, 21, 13, 15)
-                                       }
-                                },
-                                {
-                                  'o', new Character
-                                       {
-                                         Char = 'o',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(214, 0, 15, 15)
-                                       }
-                                },
-                                {
-                                  'p', new Character
-                                       {
-                                         Char = 'p',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(1, 11),
-                                         Bounds = new Rectangle(53, 0, 14, 20)
-                                       }
-                                },
-                                {
-                                  'q', new Character
-                                       {
-                                         Char = 'q',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(68, 0, 14, 20)
-                                       }
-                                },
-                                {
-                                  'r', new Character
-                                       {
-                                         Char = 'r',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 9,
-                                         Offset = new Point(1, 11),
-                                         Bounds = new Rectangle(246, 0, 9, 15)
-                                       }
-                                },
-                                {
-                                  's', new Character
-                                       {
-                                         Char = 's',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 14,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(15, 21, 14, 15)
-                                       }
-                                },
-                                {
-                                  't', new Character
-                                       {
-                                         Char = 't',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 8,
-                                         Offset = new Point(0, 6),
-                                         Bounds = new Rectangle(137, 0, 8, 20)
-                                       }
-                                },
-                                {
-                                  'u', new Character
-                                       {
-                                         Char = 'u',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 15,
-                                         Offset = new Point(1, 11),
-                                         Bounds = new Rectangle(58, 21, 13, 15)
-                                       }
-                                },
-                                {
-                                  'v', new Character
-                                       {
-                                         Char = 'v',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 13,
-                                         Offset = new Point(-1, 11),
-                                         Bounds = new Rectangle(230, 0, 15, 15)
-                                       }
-                                },
-                                {
-                                  'w', new Character
-                                       {
-                                         Char = 'w',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 19,
-                                         Offset = new Point(-1, 11),
-                                         Bounds = new Rectangle(177, 0, 20, 15)
-                                       }
-                                },
-                                {
-                                  'x', new Character
-                                       {
-                                         Char = 'x',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 12,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(86, 21, 12, 15)
-                                       }
-                                },
-                                {
-                                  'y', new Character
-                                       {
-                                         Char = 'y',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 14,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(83, 0, 14, 20)
-                                       }
-                                },
-                                {
-                                  'z', new Character
-                                       {
-                                         Char = 'z',
-                                         Channel = 15,
-                                         TexturePage = 0,
-                                         XAdvance = 13,
-                                         Offset = new Point(0, 11),
-                                         Bounds = new Rectangle(72, 21, 13, 15)
-                                       }
-                                }
+                                { 'a', new Character('a', 0, 26, 14, 15, 0, 11, 15, 0, 15) },
+                                { 'b', new Character('b', 8, 0, 14, 20, 1, 6, 15, 0, 15) },
+                                { 'c', new Character('c', 30, 21, 13, 15, 0, 11, 14, 0, 15) },
+                                { 'd', new Character('d', 23, 0, 14, 20, 0, 6, 15, 0, 15) },
+                                { 'e', new Character('e', 198, 0, 15, 15, 0, 11, 15, 0, 15) },
+                                { 'f', new Character('f', 126, 0, 10, 20, -1, 6, 7, 0, 15) },
+                                { 'g', new Character('g', 38, 0, 14, 20, 0, 11, 15, 0, 15) },
+                                { 'h', new Character('h', 98, 0, 13, 20, 1, 6, 15, 0, 15) },
+                                { 'i', new Character('i', 146, 0, 4, 20, 1, 6, 6, 0, 15) },
+                                { 'j', new Character('j', 0, 0, 7, 25, -2, 6, 6, 0, 15) },
+                                { 'k', new Character('k', 112, 0, 13, 20, 1, 6, 14, 0, 15) },
+                                { 'l', new Character('l', 151, 0, 4, 20, 1, 6, 6, 0, 15) },
+                                { 'm', new Character('m', 156, 0, 20, 15, 1, 11, 22, 0, 15) },
+                                { 'n', new Character('n', 44, 21, 13, 15, 1, 11, 15, 0, 15) },
+                                { 'o', new Character('o', 214, 0, 15, 15, 0, 11, 15, 0, 15) },
+                                { 'p', new Character('p', 53, 0, 14, 20, 1, 11, 15, 0, 15) },
+                                { 'q', new Character('q', 68, 0, 14, 20, 0, 11, 15, 0, 15) },
+                                { 'r', new Character('r', 246, 0, 9, 15, 1, 11, 9, 0, 15) },
+                                { 's', new Character('s', 15, 21, 14, 15, 0, 11, 14, 0, 15) },
+                                { 't', new Character('t', 137, 0, 8, 20, 0, 6, 8, 0, 15) },
+                                { 'u', new Character('u', 58, 21, 13, 15, 1, 11, 15, 0, 15) },
+                                { 'v', new Character('v', 230, 0, 15, 15, -1, 11, 13, 0, 15) },
+                                { 'w', new Character('w', 177, 0, 20, 15, -1, 11, 19, 0, 15) },
+                                { 'x', new Character('x', 86, 21, 12, 15, 0, 11, 12, 0, 15) },
+                                { 'y', new Character('y', 83, 0, 14, 20, 0, 11, 14, 0, 15) },
+                                { 'z', new Character('z', 72, 21, 13, 15, 0, 11, 13, 0, 15) }
                               },
-                 Charset = string.Empty
-               };
+          Charset = string.Empty
+        };
 
         return font;
       }
@@ -363,6 +106,6 @@ namespace Cyotek.Drawing.BitmapFont.Tests
       get { return Path.Combine(this.DataPath, "simple-xml.fnt"); }
     }
 
-    #endregion
+    #endregion Protected Properties
   }
 }